tests/wpt/mozilla/tests/mozilla/mozbrowser/reload.html

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {

  const SRC = "data:,foobar";
  const RELOAD_MAX_COUNT = 5;


  var iframe = document.createElement("iframe");
  iframe.mozbrowser = "true";
  iframe.src = SRC;

  var reload_count = 0;

  iframe.addEventListener("mozbrowserloadend", t.step_func(e => {
    reload_count++;
    assert_equals(SRC, e.target.src);
    if (reload_count == RELOAD_MAX_COUNT) {
      t.done();
    } else {
      iframe.reload();
    }
  }));


  document.body.appendChild(iframe);
});

</script>
</body>
